% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy metric}
\usage{
accuracy(referenceM, A)
}
\arguments{
\item{referenceM}{the matrix to use as reference}

\item{A}{the matrix obtained from a fit}
}
\description{
Compute the accuracy of a fit
}
