% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVAR.R
\name{bootstrappedVAR}
\alias{bootstrappedVAR}
\title{Bootstrap VAR}
\usage{
bootstrappedVAR(v)
}
\arguments{
\item{v}{the VAR object as from fitVAR or simulateVAR}
}
\description{
Build the bootstrapped series from the original var
}
