% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulseResponse.R
\name{checkImpulseZero}
\alias{checkImpulseZero}
\title{Check Impulse Zero}
\usage{
checkImpulseZero(irf)
}
\arguments{
\item{irf}{irf output from impulseResponse function}
}
\value{
a matrix containing the indices of the impulse response function that
are 0.
}
\description{
A function to find which entries of the impulse response function 
are zero.
}

