% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVAR.R
\name{companionVAR}
\alias{companionVAR}
\title{Companion VAR}
\usage{
companionVAR(v)
}
\arguments{
\item{v}{the VAR object as from \code{fitVAR} or \code{simulateVAR}}
}
\description{
Build the VAR(1) representation of a VAR(p) process
}
