% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVECM.R
\name{decomposePi}
\alias{decomposePi}
\title{Decompose Pi VECM matrix}
\usage{
decomposePi(vecm, rk, ...)
}
\arguments{
\item{vecm}{the VECM object}

\item{rk}{rank}

\item{...}{options for the function (TODO: specify)}
}
\value{
alpha

beta
}
\description{
A function to estimate a (possibly big) multivariate VECM time series
using penalized least squares methods, such as ENET, SCAD or MC+.
}
