% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulseResponse.R
\name{errorBandsIRF}
\alias{errorBandsIRF}
\title{Error bands for IRF}
\usage{
errorBandsIRF(v, irf, alpha, M, resampling, ...)
}
\arguments{
\item{v}{a var object as from fitVAR or simulateVAR}

\item{irf}{irf output from impulseResponse function}

\item{alpha}{level of confidence (default \code{alpha = 0.01})}

\item{M}{number of bootstrapped series (default \code{M = 100})}

\item{resampling}{type of resampling: \code{"bootstrap"} or \code{"jackknife"}}

\item{...}{some options for the estimation: \code{verbose = TRUE} or \code{FALSE}, 
\code{mode = "fast"} or \code{"slow"}, \code{threshold = TRUE} or \code{FALSE}.}
}
\value{
a matrix containing the indices of the impulse response function that
are 0.
}
\description{
A function to estimate the confidence intervals for irf and oirf.
}

