% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateVAR.R
\name{estimateVAR}
\alias{estimateVAR}
\title{Multivariate VAR estimation}
\usage{
estimateVAR(rets, penalty = "ENET", p = 1, options = NULL)
}
\arguments{
\item{rets}{the data from the time series: variables in columns and observations in
rows}

\item{penalty}{the penalty function to use. Possible values are \code{"ENET"}, \code{"SCAD"}
or \code{"MCP"}}

\item{p}{order of the VAR model (only for \code{"ENET"} penalty)}

\item{options}{options for the function (TODO: specify)}
}
\description{
A wrapper to estimate a (possibly big) multivariate VAR time series
using penalized least squares methods, such as ENET, SCAD or MC+.
}
\author{
Simone Vazzoler
}

