% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulseResponse.R
\name{impulseResponse}
\alias{impulseResponse}
\title{Impulse Response Function}
\usage{
impulseResponse(dataVar, len = 20)
}
\arguments{
\item{dataVar}{the data in the for of a VAR}

\item{len}{length of the impulse response function}
}
\value{
\code{irf} a matrix containing the impulse response function.
}
\description{
A function to estimate the Impulse Response Function of a given VAR.
}

