% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVAR.R
\name{informCrit}
\alias{informCrit}
\title{Computes information criteria for VARs}
\usage{
informCrit(v)
}
\arguments{
\item{v}{a list of VAR objects as from fitVAR.}
}
\description{
This function computes information criteria (AIC, Schwartz and
Hannan-Quinn) for VARs.
}
