% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{l2norm}
\alias{l2norm}
\title{Matrix norms}
\usage{
l2norm(M)
l1norm(M)
maxNorm(M)
frobNorm(M)
spectralRadius(M)
operatorNorm(M)
}
\arguments{
\item{M}{}
}
\description{
Some matrix norms
}
\author{
Simone Vazzoler
}

