% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcSimulations.R
\name{mcSimulations}
\alias{mcSimulations}
\title{Monte Carlo simulations}
\usage{
mcSimulations(N, nobs = 250, nMC = 100, rho = 0.5, sparsity = 0.05,
  penalty = "ENET", covariance = "toeplitz", options = NULL)
}
\arguments{
\item{N}{}

\item{nobs}{}

\item{nMC}{}

\item{rho}{}

\item{sparsity}{}

\item{penalty}{}

\item{covariance}{}

\item{options}{}
}
\value{
A list containing ...
}
\description{
This function generates monte carlo ...
}
\author{
Simone Vazzoler
}

