% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{multiplot}
\alias{multiplot}
\title{Plot VECMs}
\usage{
plotVECM(v)
}
\arguments{
\item{...}{a sequence of ggplots to be plotted in the grid.}

\item{plotlist}{a list containing ggplots as elements.}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. If present, 'cols' is ignored.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
From R Cookbook}

\item{v}{a VECM object (as from \code{fitVECM})}
}
\value{
An \code{image} plot with a specific color palette (black zero entries, red 
for the negative ones and green for the positive)

A ggplot containing the plots passed as arguments
}
\description{
Plot all the matrices of a VECM model

Multiple plot function. ggplot objects can be passed in ..., or 
to plotlist (as a list of ggplot objects)
}

