% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIRF.R
\name{plotIRF}
\alias{plotIRF}
\title{IRF plot}
\usage{
plotIRF(irf, eb, i, j, type, bands)
}
\arguments{
\item{irf}{the irf object to plot}

\item{eb}{the errorbands to plot}

\item{i}{the first index}

\item{j}{the second index}

\item{type}{\code{type = "irf"} or \code{type = "oirf"}}

\item{bands}{\code{"quantiles"} or \code{"sd"}}
}
\value{
An \code{image} plot relative to the impulse response function.
}
\description{
Plot a IRF object
}

