% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIRF.R
\name{plotIRFGrid}
\alias{plotIRFGrid}
\title{IRF grid plot}
\usage{
plotIRFGrid(irf, eb, indexes, type)
}
\arguments{
\item{irf}{the irf object computed using impulseResponse}

\item{eb}{the error bands estimated using errorBands}

\item{indexes}{a vector containing the indeces that you want to plot}

\item{type}{plot the irf (\code{type = "irf"} by default) or the orthogonal irf 
(\code{type = "oirf"})}

\item{bands}{which type of bands to plot ("quantiles" (default) or  "sd")}
}
\value{
An \code{image} plot relative to the impulse response function.
}
\description{
Plot a IRF grid object
}

