% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Matrix plot}
\usage{
plotMatrix(M, colors)
}
\arguments{
\item{M}{the matrix to plot}

\item{colors}{dark or light}
}
\value{
An \code{image} plot with a particular color palette (black zero entries, red 
for the negative ones and green for the positive)
}
\description{
Plot a sparse matrix
}
