% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotVAR}
\alias{plotVAR}
\title{VAR plot}
\usage{
plotVAR(v)
}
\arguments{
\item{v}{an object of the class sparsevar type estimate or simulation}
}
\value{
An \code{image} plot with a particular color palette (black zero entries, red 
for the negative ones and green for the positive)
}
\description{
Plot all the matrices of a VAR model
}

