% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotVECM}
\alias{plotVECM}
\title{Plot VECMs}
\usage{
plotVECM(v)
}
\arguments{
\item{v}{a VECM object (as from \code{fitVECM})}
}
\value{
An \code{image} plot with a specific color palette (black zero entries, red 
for the negative ones and green for the positive)
}
\description{
Plot all the matrices of a VECM model
}

