% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulateVAR}
\alias{simulateVAR}
\title{VAR simulation}
\usage{
simulateVAR(N = 100, nobs = 250, rho = 0.5, sparsity = 0.05,
  method = "normal", covariance = "toeplitz")
}
\arguments{
\item{N}{}

\item{nobs}{}

\item{rho}{}

\item{sparsity}{}

\item{method}{}

\item{covariance}{}
}
\value{
A list containing ...
}
\description{
This function generates a multivariate VAR time series
}
\author{
Simone Vazzoler
}

