% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar.R
\docType{package}
\name{svar}
\alias{svar}
\alias{svar-package}
\title{svar: A package to estimate multivariate time series models (such as VAR and 
VECM), under the sparsity hypothesis.}
\description{
It performs the estimation of the matrices of the models using penalized 
Least Squares methods such as LASSO, SCAD and MCP.
}
\section{svar functions}{

\code{estimateVAR}, \code{estimateVECM}
}

