% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVAR.R
\name{testGranger}
\alias{testGranger}
\title{Test for Ganger Causality}
\usage{
testGranger(v, eb)
}
\arguments{
\item{v}{the VAR object as from fitVAR or simulateVAR}

\item{eb}{the error bands as obtained from errorBands}
}
\description{
This function should retain only the coefficients of the 
matrices of the VAR that are statistically significative (from the bootstrap)
}

