% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVAR.R
\name{transformData}
\alias{transformData}
\title{Transorm data}
\usage{
transformData(data, p, opt)
}
\arguments{
\item{data}{the data}

\item{p}{the order of the VAR}

\item{opt}{a list containing the options}
}
\description{
Transform the input data
}
