% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVAR.R
\name{varENET}
\alias{varENET}
\title{VAR ENET}
\usage{
varENET(data, p, lambdas, opt)
}
\arguments{
\item{data}{the data}

\item{p}{the order of the VAR}

\item{lambdas}{a vector containing the lambdas to be used in the fit}

\item{opt}{a list containing the options}
}
\description{
Estimate VAR using ENET penalty
}
