% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVAR.R
\name{computeForecasts}
\alias{computeForecasts}
\title{Computes forecasts for VARs}
\usage{
computeForecasts(v, numSteps)
}
\arguments{
\item{v}{a VAR object as from fitVAR.}

\item{numSteps}{the number of forecasts to produce.}
}
\description{
This function computes forecasts for a given VAR.
}

