% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVAR.R
\name{estimateVAR}
\alias{estimateVAR}
\title{Multivariate VAR estimation}
\usage{
estimateVAR(data, p = 1, penalty = "ENET", options = NULL)
}
\arguments{
\item{\code{data}}{the data from the time series: variables in columns and observations in 
rows}

\item{\code{p}}{order of the VAR model}

\item{\code{penalty}}{the penalty function to use. Possible values are \code{"ENET"}, 
\code{"SCAD"} or \code{"MCP"}}

\item{\code{options}}{options for the function (TODO: specify)}
}
\value{
\code{A} the list (of length \code{p}) of the estimated matrices of the process

\code{fit} the results of the penalized LS estimation

\code{mse} the mean square error of the cross validation

\code{time} elapsed time for the estimation
}
\description{
A function to estimate a (possibly big) multivariate VAR time series
using penalized least squares methods, such as ENET, SCAD or MC+.
}
\author{
Simone Vazzoler
}

