% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l2norm}
\alias{l2norm}
\title{Matrix norms}
\usage{
\code{l2norm(M)}
\code{l1norm(M)}
\code{lInftyNorm(M)}
\code{maxNorm(M)}
\code{frobNorm(M)}
\code{spectralRadius(M)}
\code{spectralNorm(M)}
}
\arguments{
\item{M}{the matrix (real or complex valued)}
}
\description{
Some matrix norms
}
\author{
Simone Vazzoler
}

