% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIRF.R
\name{plotIRF}
\alias{plotIRF}
\title{IRF plot}
\usage{
plotIRF(M)
}
\arguments{
\item{irf}{the matrix to plot}
}
\value{
An \code{image} plot with a particular color palette (black zero entries, red 
for the negative ones and green for the positive)
}
\description{
Plot a IRF object
}

