% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Plot the matrix}
\usage{
plotMatrix(M)
}
\arguments{
\item{M}{the matrix to plot}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Plot the matrix
}
\examples{
add(1, 1)
add(10, 1)
}

