% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Matrix plot}
\usage{
plotMatrix(M)
}
\arguments{
\item{M}{the matrix to plot}
}
\value{
An \code{image} plot with a particular color palette
}
\description{
Plot a sparse matrix
}
\examples{
plotMatrix(M)
}
\author{
Simone Vazzoler
}

