% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotVAR}
\alias{plotVAR}
\title{Plot VARs}
\usage{
plotVAR(..., colors)
}
\arguments{
\item{...}{a sequence of VAR objects (one or more
than one, as from \code{simulateVAR} or \code{fitVAR})}

\item{colors}{the gradient used to plot the matrix. It can be "light" (low =
red -- mid = white -- high = blue) or "dark" (low = red -- mid = black -- 
high = green)}
}
\value{
An \code{image} plot with a specific color palette
}
\description{
Plot all the matrices of a VAR model
}
