% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotVAR}
\alias{plotVAR}
\title{Plot VARs}
\usage{
plotVAR(...)
}
\arguments{
\item{...}{a sequence of VAR objects (one or more
than one, as from \code{simulateVAR} or \code{fitVAR})}
}
\value{
An \code{image} plot with a specific color palette (black zero entries, red 
for the negative ones and green for the positive)
}
\description{
Plot all the matrices of a VAR model
}

