% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevar.R
\docType{package}
\name{svar}
\alias{svar}
\alias{svar-package}
\title{sparsevar: A package to estimate multivariate time series models (such as VAR and 
VECM), under the sparsity hypothesis.}
\description{
It performs the estimation of the matrices of the models using penalized 
least squares methods such as LASSO, SCAD and MCP.
}
\section{svar functions}{

\code{estimateVAR}, \code{estimateVECM}, \code{simulateVAR}, \code{createSparseMatrix},
\code{mcSimulations}, \code{plotMatrix}, \code{plotVAR}, \code{plotComparisonVAR},
\code{l2norm}, \code{l1norm}, \code{lInftyNorm}, \code{maxNorm}, \code{frobNorm},
\code{spectralRadius}, \code{spectralNorm}
}

